backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
pymongo==4.6.0
flask-jwt-extended==4.6.0
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
flask-cors==4.0.0
cryptography==43.0.1